pyproject.toml
[project]
name = "t-i"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "t-tech-investments>=0.3.5",
]
